Refector: categoryId Long으로 수정

+++ src/main/kotlin/com/hwalaon/wezxro_server/domain/service/controller/ServiceAdminController.kt
@@ -19,7 +19,7 @@ class ServiceAdminController(
     @GetMapping("/list")
     fun getServices(
         @AuthenticationPrincipal userInfo: PrincipalDetails,
-        @RequestParam(name = "category", defaultValue = "0") categoryId: Int) =
+        @RequestParam(name = "category", defaultValue = "0") categoryId: Long) =
         queryAdminServiceService.serviceList(userInfo.account.clientId, categoryId).let {
             BasicResponse.ok(it)
         }
